feat: done module 1
